backendaiapi/requirements.txt
Flask
Flask-Cors
langchain
langchain-google-genai
python-dotenv